Specify a pattern for the sign of numbers.
For integers and decimal numbers (reals), allow the user to specify the
pattern of the signs (leading sign, and for reals, the sign of the exponent).
This gives the user the option to ask for a pattern that matches unsigned
numbers (by specifying the empty string as the pattern).
Requested by "Wilson, Jonathan" <jonathan.wilson@Vanderbilt.Edu>.