Wrap the delimited clauses inside a (?|)
We now wrap the clauses for delimited patterns (and quoted) inside a (?|).
This has the advantage that the captures under -keep are always in
$1, $2, $3, and $4, and no other capture variables will be set, regardless
of how many delimiters will be used.

It will require the use of 5.10.0 (and hence, the patterns will no longer
be available for older perls).

The test file, t/test_delimited.t was complete rewritten, and now uses
Test::Regexp to run the tests.