Dropping support for pre 5.10.0 perls.
It becomes too painful to support older perls; it's too easy to
write constructs which have been in use for the past 8 years, to
have them fail on old versions of perl.

If you are running an older perl, and really need the newest versions
of Regexp::Common, please consider upgrading your perl.