Disable leading 0s being optional by default.
It was once argued that for Australian postal codes which start with a 0,
the leading 0 is optional, and the regular expression supported both this.
But there doesn't seem to be any evidence that this is actually the case.
So, from now on, we alway require Australian postal codes to be 4 digits,
even if the first digit is a 0.

The old behaviour, with an optional leading zero, can be gotten by using
the {-lax} parameter.