Improve reporting requirements.
Now that we dropped support for pre-5.10.0, we also worked around a
missing warnings.pm, so we have to require it.

Also, we're using Config in one of the files; might as well make it
a prerequisite, although it's unlikely to be missing, as it's an
essential part of the main distro.