Update pattern for MySQL comments.
MySQL no longer seems to allow comments of the form

    /* Comment ;

(if it ever did).

We now no longer accept such comments.

We also relaxed the rules for '--' style MySQL comments. Previously, we
required the '--' token to be followed by a space. Now we allow any whitespace.
In particular, this makes "--\n" valid, where it previously did not.